Fix Sprite Sheet Timeline out of memory exception
Calling Preload create a graphic device that eventually cause a crash.
Since timelines are copied a lot for the Undo fonctionnality it doesn't
take a lot of time to run out of memory.

It also doesn't need to be loaded as it's only used in the getter that
create its own SpriteSheetHelperDialog.