Update BaseSkillActivation to prevent double activation when checking
The old way to use skills was to check availability and then use the skill,
which would cause another check. If there was a requirement using a random
number, the second check could fail even if the first one worked.
Now the first check return the successful activation so they can be used
directly without a second check.